requirements.txt
# Core dependencies
notion-client>=2.3.0
python-dotenv>=1.0.0
# OCR dependencies removed - using LLM-only approach
# pytesseract>=0.3.13
# pdf2image>=1.17.0
# pdfplumber>=0.11.6
reportlab>=4.0.0
ollama>=0.1.6

# Settings and configuration
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Error handling and retries
tenacity>=8.2.0

# Logging and output
structlog>=24.1.0
rich>=13.7.0

# Development dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-dotenv>=0.5.2
coverage>=7.8.0
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0